+++ crates/lib/docs_rs_crate_zip/src/lib.rs
@@ -6,9 +6,7 @@
 //! Also we copied the manifest structs from there.
 
 mod error;
-mod manifest;
 mod source_archive;
 
 pub use error::Error;
-pub use manifest::{FileEntry, Manifest};
-pub use source_archive::SourceArchive;
+pub use source_archive::{FileEntry, Manifest, SourceArchive};
