+++ crates/lib/docs_rs_crate_zip/src/source_archive.rs
@@ -1,7 +1,4 @@
-use crate::{
-    error::Error,
-    manifest::{FileEntry, Manifest},
-};
+use crate::error::Error;
 use async_compression::tokio::bufread::DeflateDecoder;
 use docs_rs_utils::APP_USER_AGENT;
 use futures_util::TryStreamExt as _;
@@ -9,8 +6,34 @@ use reqwest::{
     IntoUrl, StatusCode, Url,
     header::{HeaderValue, RANGE, USER_AGENT},
 };
+use serde::{Deserialize, Serialize};
 use tokio::io::{self, AsyncWrite, AsyncWriteExt as _};
 use tokio_util::io::StreamReader;
+
+/// archive manifest serde structs, copied from
+/// https://github.com/rust-lang/crates.io/blob/5274087feb193ee490e9a6bbdf2e18e74e9ddaeb/crates/crates_io_crate_zip/src/lib.rs
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct Manifest {
+    /// One entry per file in the zip, sorted alphabetically by path.
+    pub files: Vec<FileEntry>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub struct FileEntry {
+    /// Realtive path (without the leading `{name}-{version}/` component of
+    /// the tarball).
+    pub path: String,
+    /// Byte offset in the zip where this entry's compressed payload begins.
+    pub data_offset: u64,
+    /// Length of the compressed contents in bytes.
+    pub compressed_size: u64,
+    /// Length of the uncompressed contents in bytes.
+    pub uncompressed_size: u64,
+    /// How the payload is compressed: `"deflate"` or `"store"`.
+    pub compression: String,
+    /// Lowercase hex sha256 of the uncompressed contents.
+    pub sha256: String,
+}
 
 pub struct SourceArchive {
     manifest: Manifest,
